Initial commit
In 1970, Dr. Winston Roye described a dysfunctional model which became
known as the classical waterfall model. What defined the model is the
idea of strictly separated development phases which are passed in order.
However, in which phases the overall development (and maintenance) is
divided varies between various descriptions of the model. Roye himself
described the following phases:

 * System requirements
 * Software requirements
 * Analysis
 * Program design
 * Coding
 * Testing
 * Operations

The model has a number of problems. For example, it is said to function
particularly bad when requirements are only vaguely defined or the
development is the subject of late change-requests, e.g. if the customer
doesn't know that they want --that's why everybody and their dog tries
to do "agile" development, often doing it wrong and/or without
considering whether it's actually the right choice for the project.

Another problem which often surfaces is a high number of defects which
are only discovered during the testing phase, that is, late in the
development (albeit depending on the exact definition, unit-tests and
"proving" are part of the coding phase). Since the later a given defect
is discovered the higher the effort necessary for the fix [citation
needed], this will contribute significantly to the overall development
cost.

Now, proponents of strongly typed languages (such as myself) will
postulate that a good type system will generally help discover defects
early (even quasi-instantaneously from a practical point of view), as
we effectively have the compiler will reject malformed software. Of
course, active constructive measurements taken by the designer or
developer play an important role in whether the type-system is actually
utilized to this end or not.

As for surprisingly many aspects of software-engineering methodology,
good statistical data is apparently hard to come by or doesn't actually
exist, and existing studies usually come with various strings attached.
If type systems do facilitate early detection of defects, the situation
should have improved significantly in the last 50 years. We'll test this
by trying to program a small multiplayer game in Rust using a strict
waterfall process using the following phases:

 * Software requirements: we define the game mechanics and user
   interface, as detailed as feasible.
 * Analysis/Design: we define the modules and their scope.
 * Implementation: we implement each module. Compiling will be allowed,
   but running the program or any other form of dynamic test will not.
 * Testing: we write unit tests for the individual modules, run them,
   and fix any defect we find. However, we do not run the program itself
   or perform any test which would involve any form of user-interaction.
 * Operation: we do what game publisher do nowadays and let our
   customers find any remaining issues.

We'll keep a trail of the development, especially all the issues we may
encounter during development, using git. As for the a game, fittingly
for 2020 we choose to make a variant of "Dr. Mario", a NES game.
Hopefully, the game will be finished in time for new year's eve.