spec: initial version
We deem the specification sufficiently complete at this point. Now we
wait for comments (we already received a few) and incorporate them into
the specification.

We initially planned on also specifying the command options but decided
against it. Rather, we'll prepare a section for the README after the
spec phase ended, on a separate branch which will be merged late in the
implementation phase.