readme: add (admin) usage information
We left the CLI interface out of the actual specification. However, we
still do want to define it before the implementation phase. Thus, we
decided to prepare this snippet of the README, in the hope that we
covered every option necessary for operations.

Note that we did _not_ include any option for controling logging or
debugging. Since this is a case-study, we'll make a point of living a
dangerous life, with only the most critical errors being reported, right
before aborting the program.