design: initial version
Originally we intended to finish the game in time for new year's eve.
Now it's the 6th of January, I'll have to work tomorrow, again, and we
just finished the (preliminary) design. In part I blame this on being
pre-occupied with other stuff as well as a severe lack of motivation.
However, elaborating the design in text also took much more time than
expected, and lead to some questions I had to answer. Which is actually
a good thing, since those had turned up sooner or later, anyway.

I'll reiterate over the design as soon as I have some time and
motivation, again. After that, we'll start the implementation phase,
which hopefully will go smoothly, considering how much time we spent in
the design phase.