End of analysis/design phase
We'll close the design phase. Thus, the implementation phase begins now.
However, we don't know yet when we'll actually start implementing the
design. Hopefully, we covered everything important in the design and
don't encounter anything unforseen.

The previous fixes applied during the design phase also include one for
the spec. Hoever, this trivial fix did affect the design in any way.

A bigger issue we already ran into would be our poor choice of trailers
used for marking fixes. Rather than declaring the phase, which is
trivial to determine from the commit's "position" in the graph alone, we
really should have included some form of categorization, e.g. functional
changes vs. typos, or the severeness. Changing it now would require
rewriting most of the history, so we're keeping the old trailers. For
the sake of consistency, we'll also _keep_ using them for future fixes.