Add basic rust template with dependencies
... and a tokio-based main function

In addition to `clap`, `rand_pcg` and `tokio`, which were mentioned in
the design documents, we pull in some of their dependencies explicitly
since we'll need to refer to some of them in our code (e.g. import a
trait or type). We also pull in `log`, which was not mentioned in the
design document. Obviously, we _do_ want to log at least warnings and
errors. Using `log` makes this a lot more convenient.