design: allow elements to settle in top row, transiently
We previously neglected that an element may settle, but be eliminated
right after. We do want to allow this. Therefore, we need to make the
top row accessible in the field of moving elements.

The specification appears to be rather vague in this regard, so it's ok
to change this in the design only.