util: add types for position and direction
We'll often use a `ColumnIndex` and `RowIndex` together as a position.
This type improves readability in those cases and faciliates a
convention.

The direction will be important especially for linking capsule elements
together. Computing the position relative to anothe position will also
be a common task. We therefore supply an impl for `Add`.