field: add convenience functions to TileContents
We'll need to perform specific operations of contents when implementing
some gameplay logic in the future, which would normally clutter our code
with `match`es. These functions should cover all relevant cases, saving
us from said clutter.

Most of these cover different accesses to an inner capsule element. This
is motivated by the `partner` member of the `CapsuleElement` type, which
will be the target of both queries and mutations. This is not the case
for `Virus`es, neither is it for free tiles.