util: add a utility for iterating over a row
We'll often want to iterate over all elements in a specific row. In such
cases, we'd need to transform `ColumnIndex`es to `Position`s. While we
don't need much code to do so, this would cause quite a bit of
repetition in some modules. Thus, we decided to encapsulate this mapping
into a funciton.