gameplay: use a dedicated return type for static_field::row_of_four
Previously, we'd collect the positions of elements in such rows in a
`Vec`. Given the nature of these rows, however, they can also be
repsesented by a range for one component of the positions and a fixed
value for the other. The resulting type appears more complex than a
`Vec`, but the runtime cost should be significantly lower.