gameplay: make row_of_four generic
When introducing the `row_of_four` function originally in

        83af7c9c95bbc65146523d219ca4d4bdbe38715f

we overlooked the major requirement for it to be based on `Index` and
`PotentiallyColoured` rather than a concrete field type. It ended up
errorneously depending on `StaticField` and in the wrong location.

This change moves the function into the correct module, generalizing it.
The changes related to the generalization happen to be restricted to the
function signature.