display: replace Element trait with ElementFactory trait
The `Element` trait was introduced as a means for the placement of
display elements via methods of `Area`. For the existing elements, the
trait served its purpose as the width and height of those elements was
statically defined. However, we'll also need to place differently sized
text elements, and we don't want to define a type for each of those.

Hence, we decided to switch from a trait implemented for the display
elements themselves to a trait for factories of those display elements.
This will still allow querying the size of the elements to be created
beforehand but also allow that size to depend on data which will be
transferred to the element upon creation.

We will not classify this change as a bug-fix since the entire `Area`
concept was not defined during the design phase anyway.