game: make waiting phase function generic over entire phase-end message
While the phase end is recognized within `waiting`, the phase end
message itself is processed outside the scope of the funciton. Hence,
the message does not need to actually know anything about this
particular message type. And neither will the the phase's control
function, since it won't be responsible for sending the phase end
message.

As a result, we also remove the completely useless local `PhaseEnd`
type.