game: remove unused EndData type from waiting
Both transmission and processing of the phase end message will be
outside the scope of the phase and phase control functions. In fact, the
message's contents doesn't have anything to do with the waiting phase
itself. Hence, having a declaration related to this type being part of
the module is unreasonable.

Rather than moving it to the toplevel `game` module, we choose to simply
remove the type for now, since we may make different decisions regaring
its design when actually needing to define it.