game: remove some residuals from Actor::new
These residuals were leftovers from when we constructed the function
originally over an extended period of time. They ended up being
committed in

        90cfbb5b4ce4d220d94e6f82294a10f0dd9d95e5

The extra `next_colours` deserves some attention since it shadowed the
parameter of the same name. Thus, the first capsule in a round would
always be red. Which is not random. I.e. we categorize this as an actual
bug which we cought early, and only because we skimmed through the many
"unused" warnings which accululated up to this point.

Fix-phase: implementation