game: extract capsule receiver from scoreboard round::round
... rather than just getting a lock.

Holding a lock, which is `!Send` over a suspension point (an `await`)
prevents us from using the function in a spawned context. Thus, we have
to choose another approach for keeping the capsule receiver.

Fortunately, the receiver is only used from the round phase function (or
rather, a single `Actor` instance) anyway. Hence, we can as well "just"
take the receiver's ownership. In order to do so, we need to be able to
do the "take" part. Hence, we wrap the receiver in an `Option`.