game: keep peer's address in the connection token
Roster entries need to be initialized with the peer's socket address.
We'll get the address when accepting a connection and we'll need to keep
the address around until the player is registered.

We already keep around a connection token which is intended as an
identification for registration. Hence, associating the address with the
conneciton token appears natural.