End of analysis/design phase, take 3
We ended up making more extensive changes to the design than last time.
This is likely the case because we used `tokio` for the very first time
and hit some unexpected problems, even though we had a thorough look at
the documentaiton while writing the initial design.

We'll salvage a lot of the commits from the previous implementation
phase. It we're lucky, we'll be finished soon (famous last words).