error: add extension trait for more convenient reporting
Especially in tasks which will serve a single conneciton, error handling
will, more or less, be restricted to logging the error and otherwise
ignoring it. However, doing so in explicit `map`s will add significant
and repetetive boiler-plate. Hence, we abstract it away behind an
interface which we apply to the two most prominent implementations of
`Try`.