error: add convenience Error wrapping Debug type
Many functions related to synchronization and data transport mechanisms
return a `Result` with error value types which do not implement `Error`.
For instance, in the case of a "send" or "update" function, `Err` may
hold the value the user tried to send or update. However, virtually all
relevant types do implement `Debug`.