deps: pull in rand rather than rand_core
In the past, we shied away from pulling in `rand`, since we deemed it a
somewhat costly dependency, and the same could be achieved with
`rand_core` only. However, it _does_ provide a few interfaces which make
generating random values far more ergonomic. Still, we choose to restrict
which features we're going to pull in. In particular, we don't pull in
the thread-local RNG as we really only want to use deterministig PRNGs.

Of cource, we could deploy our own convenience-APIs, but that would be
just silly.