field: restructure tile initialization in prepare_field
I realized that with a bit of reordering, we can write the
initialization more concise.

These changes actually contain an important fix: previously, we'd select
the position using a `take` followed by a `last`. However, the way the
`take` was written, we'd allow it to return `0` items.

Now, we use `nth` for selecting the field, which should have been the
obvious thing to do, anyway.