game: add utility for awaiting a transition
The design document mentions "type erasure via a trait". That wording is
sub-optimal as it suggests that we'll define a trait exposing an `async`
function (which is not possible) when the intention was to only hide
some of the detection logic behind a trait.

Ultimately, we decided to simply use `Fn(&P) -> bool` rather than
defining a named trait.