game: introduce special error type for connection task functions
We already introduced a bunch of error types in the dedicated `error`
module, originally with the inention to omit module-specific error
types.

However, with connection tasks we want to special-case (presumably)
voluntary termination of a player's connection. These are not errors in
the specific sense: we are absolutely satisfied with the information
that a player disconnected and don't need to know the specific how or
why. We also want this difference to be reflected in the reporting both
to the player in question and the logs.

Up until now, we (ab)used `std::io::error::ErrorKind::UnexpectedEof` to
express such a condition. However, we will also have to handle error
types other than `std::io::Error`. This requires boxing or wrapping
errors in some other way, which makes recognizing this error kind
unreasonably complex.

Instead, we decided to go the other way round and have the condition
expressed via the "top-level" error type, i.e. have an error type which
either expresses the termination condition or carries another error. For
flexibility, we box the inner error.

Sadly, we cannot simply implement `From<impl std::error::Error>` for
this type, as this would clash with the blanked `Into` impl (and I
wonder how this actually works for `Box`). As a significant portion of
potential sources do produce I/O errors, we implement at least `From`
for types convertible to those. In addition, we introduce a fn for
encapsulating errors which should work well with `Result::map_err`.