player: make Roster a Vec<Tag> rather than a Vec<Arc<Data>>
When introducing `Roster` in

        d483aa219344726b5f69da03dba93dd4c2055ef9

we didn't give it too much thought, but using `Arc<Data>` seemed like a
resonable choice given that it would allow both comparison with `Tag`
and reading player data.

However, when introducing the `ScoreBoard` display entity and the
`Entry` trait in

        96deb9c81204e5d87784b36236224cd0317996a2

We decided to make all `Entry` provide access to player data via a `Tag`
rather than an `Arc<Data>`. This created a problem since, on purpose, we
didn't provide any means to create a `Tag` from an `Arc<Data>`. Hence,
we either needed to rewrite the score board entry to provde an
`Arc<Data>` rather than a `Tag` or make `Roster` hold `Tag`s.

We decided on the latter, not just because it's the smaller change but
also because a `Tag` is more convenient to have.