game: add types encapsulating channels for the round phase
We want to hide phase specific internals from the outside. However, we
do need to handle the channels in the toplevel `game` module. The design
specification already mentions a type encapsulating the senders and
receivers used by the connection task in an opaque struct. We decided to
do the same for the control task.