game: introduce special casing of WouldBlock I/O errors
We originally introduced ConnTaskError to allow transportation of
arbitrary errors, i.e. types other than `std::io::Error` and properly
represent the special connection termination at the same time.

One `std::io::ErrorKind` we forgot about was `WouldBlock`, which we
usually want to specialcase. If we still want to do so, we need to
expose this condition from a `ConnTaskError`. The sanest way to do so is
using an additional variant.