console: add fn implementing game master logic
At some point we intended to conditionally also serve a GM console via
stdin and stdout. However, using those as streams comes with a few
rather nasty caveats, so we decide to drop this plan for now. We get
away with this because we never mentioned it in the spec.