main: add initial version of main
We finally reached the final integration step. Now, we _could_ run the
game but we'll refrain from doing so until the very end of the testing
phase.

We will not deive into that phase right away either, but have a look at
the remaining warnings now that a build doesn't produce an endless trail
of "unused xyz"-warnings. We already spottet one suspicious warning
rregarding something in `game::round`.

We also want to perform a little cleanup --but no changes which would
violate the specification. We also want to make use of the disconnect
notification mechanism we introduced.