field: add trait for items which may be capsule elements
This trait could have come in handy during the implementation phase, and
it does mirror a function of `TileContents`, but apparently we did do
well enough without it.

However, we really want this for testing. In some tests, we want to make
sure operations don't break existing capsule partnerships, neither in
static nor dynamic fields. Obviously, we don't want to duplicate the
logic but formulate the test on a common interface.

This trait is the missing piece for that interface.