field: avoid endless loop in preparation::prepare_field
Previously, a part of `preparation::prepare_field` was based on the
assumption that, in a partially filled field free of rows-of-four
configuration, colouring any non-occupied tile would lead to a
row-of-four with at most two of the existing three colours.

For example, in a configuration like the following:

      B
    RRXR
      B
      B

colouring the tile X either red or blue would result in a row-of-four
configuration while colouring it yellow would not. However,
configurations exist where all of the three colours would result in a
row-of-four configuration. An obvious example would be:

    RRRXBBB
       Y
       Y
       Y

Therefore, our initial assumption does not hold.

We choose to leave the tile unoccupied in such situations. The
function's documentation does not need any adjustment as it already
advertises "at most" some amount of viruses. As the probability of
encountering a situation is quite low, ending up with one virus short of
the configured amount was deemed acceptable.

Fix-phase: testing