game: make sure score board entries are not left with a round score of 0
Previously, we'd initialize `round::ScoreBoardEntry`s with a round score
of `0`. Since the round score would only be updated upon the elimination
of viruses, a player could have "won" a round by making sure not to
eliminate any virus while avoiding defeat.

This change makes sure the score board entries for a round are
initialized with the initial number of viruses, as intended.

Fix-phase: testing