field: add type for aiding the construction of partially occupied fields
For testing (un)settling, we really want partially occupied fields which
more or less reflect a situation we may encounter during gameplay. Some
of our existing may also benefit from a more realistic field setup.

During a game, fields will be occupied with capsules and elements which
will be introduced one after another. We choose an approach which kinda
simulates this process: we generate a sequence of capsules (or elements)
independently from another, with each one represented by instances of
the type introduced by this change. This, of course, introduces the
potential for conflicts, which we resolve during placement of the
capsules' elements by omitting placement of elements on any field which
is already occupied.

Future changes will introduce helpers for generating complete fields
based on this helper.