field: make prepare_field safe for number_of_virs > availible area
Previously, `prepare_field` would panic when called with a desired
number of viruses greater than the number of tiles in the area availible
for virus placement.

This change avoids the problem by bailing out early if the area is
exhaused. This is, of course, less efficient than bounding the number of
viruses before starting to place viruses. However, we still deem it
acceptable as `number_of_virs` can be at most `255` due to its type.