field: add type for constructing partally occupied moving fields
For testing (un)settling, we really want partially occupied fields which
more or less reflect a situation we may encounter during gameplay. Some
of our existing may also benefit from a more realistic field setup.

During a game, fields will be occupied with capsules and elements which
will be introduced one after another. We choose an approach which kinda
simulates this process: we generate a sequence of capsules (or elements)
independently from another. We recently introduced a type for generating
these capsules.

This change introduces a utility for initializing a moving field from a
set of randomly generated capsules.