util: test iteration over RangeInclusive both forwards and backwards
We recently found an issue with reverse iteration over a range --in a
currently uncommitted test in a layer further up. Testing iterators sure
makes sense it it was kinda dumb of us to omit these previously.

Still, we only test the implementation against rows, not columns. Should
be sufficient, right?