field: impl From<Vec<Position>> for Settled in test builds
We want to test `tick::eliminate_elements`, which takes a `Settled` as
an input parameter. We could have implemented `quickcheck::Arbitrary`
for `Settled` instead, as we plan to formulate those tests using
`quickcheck`, anyway. However, this would require alos implementing
`Clone` and `Debug`.

Thus, we decided to implement this simple convertion, and only for test
builds.