field: impl From<HashSet> for Eliminated in test builds
We want to test `tick::unsettle_elements`, which takes a `Eliminated` as
an input parameter. We could have implemented `quickcheck::Arbitrary`
for `Eliminated` instead, as we plan to formulate those tests using
`quickcheck`, anyway. However, this would require alos implementing
`Clone` and `Debug`.

Thus, we decided to implement this simple convertion, and only for test
builds.