player: add utility for generting a valid player name
We plan to introduce a few tests to the `player` module. For some of
them, we'll need to initialize some `player::Data` with meaningful data,
including a player name.

We will probably not rely on an actually valid name for any of the tests
in the module itself. However, tests we plan to introduce to other
modules may well depend on these names to be valid.