display: add fn for decoding DrawCommands in tests
Our display code might end up being a pit of subtile errors.
Unfortunately, writing tests for it will take some effort.

Most of the code actually producing output pushes `DrawCommand`s through
a `DisplayHandle`, which wraps a `FramedWrite` with our `ANSIEncoder`.
If we could generalize the handle over `AsyncWrite`s with `DrawCommand`
as item type, we could simply take those and process them in our tests.
However, the whole purpose of `DisplayHandle` was to hide `DrawCommand`
from the outside.

We could somehow replace it with some mock in our tests, but that would
be cheating, and messy, and somehow not right. Instead, we take the
detour over the actual bytes we'd normally send to the player via the
network.

This obviously means we need to _decode_ `DrawCommand`s.

The obvious solution would be to write an `ANSIDecoder` and have it impl
`tokio_util::codec::Decoder`. However, we want to keep things as simple
as possible, and we'll want the ability to perform the decoding in a
specialized `AsyncWrite` we'll introduce for our tests without too much
of a headache.

So we introduce a single, "simple" decoding function. It will not cope
with cut-off ANSI sequences, but that shouldn't be a problem since we
expect `FramedWrite` to push only complete sequences anyway. If not,
we'll figure something out...