display: fix splitting of subareas from an Area
In `Area::split_top` and `Area::split_left`, we split off an `Area`
which borrows the current `Area`'s `handle`. We do this by resizing the
current area and contstructing a new `Area` covering the freed area of
the display.

When doing so, we previously made the error of using geometry
information from the current `Area` which was previously altered as part
of the resizing process. This change resolves the issue by storing the
data in question before altering it and using it to construct the new
`Area`.

The issue fixed by this change was discovered while writing tests, which
will be introduced by future commits.

Fix-phase: testing