display: add fn for creating an Area at a given position for tests
For tests, we want to generate random `Area`s, i.e. ones placed
arbitrarily on an area bounded by the `u16` type used for rows and
columns. Outside the submodule, the only possibility availible was the
`create_area` function, which allows supplying a size but no origin
position. These means are not sufficient for tests which reside outside
the submodule.

As we weant to generate random `Area`s, we could also implement
`quickcheck::Arbitrary` for `Area` if this wouldn't require `Area` to
also implement `Clone` and `Debug`.

To allow generation of random `Area`s, we introduce an additional
creation method which will only be availible in test builds.