display: fix bounding of upper/left bound in pad_top/pad_left
When padding an `Area`, we must make sure that the type's invariants are
not violated: `row_a <= row_b` and `col_a <= col_b`. We did this by
bounding one with the other. However, in `pad_top` and `pad_left` we
made a classical C&P-error, which resulted in faulty bounding.

The issue fixed by this change was discovered while writing tests, which
will be introduced by future commits.

Fix-phase: testing