display: remove unused fn Display::into_writer
In order to fix a broken RAII pattern implementation for the
`DisplayHandle` type, we need to make some changes which will involve
how the `tokio::io::AsyncWrite` used in `Display` is stored. We could
redefine this function if the time comes, but we don't use it anywhere,
anyway. We therefore decided to just remove it in advance.