deps: upgrade tokio to 1.15
In order to fix a broken RAII pattern implementation for the
`DisplayHandle` type, we need to make some changes. We plan to have
`Display` take ownership of the underlying writer, wrapped in a
`tokio::sync::Mutex`.

In some of our tests, we dissaminate the output produced by our
display code collected in a `Vec` supplied as underlying write.
Currently, we benefit from a blanked impl of the important write trait
for mutable references. However, in the future we'll need to move the
`Vec` inside the mutex and thus we'll need to access the output via that
mutex.

Normally, this would require setting up and handling an asynchronous
context. However, `tokio::sync::Mutex` gained a fn `blocking_lock` which
is usable outside an asynchronous context. We upgrade tokio in order to
use this function in our testing code.