display: migrate DrawHandle to new underlying mechanism
`DrawHandle` is meant to function as an RAII handle which send out a
termination sequence when dropped. This obviously involves I/O and thus
asynchronous work in our current overall design. As there is no such
thing as an asynchronous `Drop` (yet?), we needed to make this blocking
somehow. We did so calling `tokio::runtime::Runtime::block_on` on a
newly created runtime.

If we would have read the documentation more carefully, it would have
known that this doesn't work.

Currently, the only possibility to kick off asynchronous work from a
non-async function appears to be via newly-spawned tasks, which requires
the underlying `tokio::io::AsyncWrite` to be `Send`, `'static` and
mutably accessible from the new context.

We thus redesigned the innards of `DrawHandle` (and by extension some of
`Display`) to encapsulate an `tokio::sync::OwnedMutexGuard` which we
transfer to a new context responsible for sending the termination
sequence when dropping the handle. This way we make sure all commands
send over a handle are atomic, including the termination sequence.

These changes also cause quite a few changes in tests as we'd
unknowingly rely on them _not_ being run from an async context when now
we _do_ need handles to be created from such a context. In addition, we
needed to rewrite the `draw_handle_drop` test due to the new `'static`
requirement on the underlying `tokio::io::AsyncWrite`.

Fix-phase: testing