player: prevent tests::Name from starting or ending with whitespace
Previously, we were a lot more permissive than we ought to be: player
names should always be trimmed. This is technically more strict than the
specification. However, it doesn't go directly gainst the specification
and can be arguted to be within its spirit. It also assures that
nusernames are distinguishable for (human) players.

This change makes sure the `tests::Name` helper for generating player
names conforms to this property. In addition, we introduce a helper
function which we make public for use in future tests, possibly in other
modules.

We may also need to introduce this constraint to actual user
registration code in future commits.