game: trim player name retrieved from line input field
We require player names to be unique. We should also make sure that they
are distinguishable by end users. Names ending with whitespaces open up
an opportunity for names which are undistinguishable on the score board.
Hence, we should remove trailing whitespace. While we're at it, we may
as well remote any _preceeding_ whitespace.