game: add test-specific fn for lobby::ControlPorts
We introduced the various `Ports` and `ControlPorts` structs in order to
encapsulate and hide some phase-specific types from outside modules.
However, we choose to situate tests for phase-specific behaviour in
`gmae::tests` which is a sibling of `game::lobby` and thus doesn't has
access to its internals. Thus we'll need some help from within the
`lobby` module.