game: loop only until phase transition
Previously, we'd depend on the loop being broken upon receival of a
phase update. Testing indicates this is not always as robust as we'd
liked, resulting in a connection task being potentially stuck in the
waiting phase.

This issue was caught by a test which was is not yet comitted.

Fix-phase: testing