player: add utility for generating Handles for tests
We originally decided against implementing `quickcheck::Arbitrary` for
`Handle` since that would require implementing `Clone` for that type,
which we definitely don't want to. That point still stands, but we
discovered that we _do_ want some possibility for generting them in an
easy fashion, without constructing them from names and addresses
generated separately.