field: add fn for making capsules consistent to tests::RandomCapsules
For tests involving prepared fields, we generate a set of
`RandomCapsules`. Currently, we'd accept inconsistencies in those sets
during generation and filter them out during the transformation into an
actual `field::StaticField` or `field::MovingField`.

This approach turned out to be non-ideal. One of the pain-points it
introduces was reports of test failures listing a different set of
capsule elements than actually tested against. In addition, it makes
introducing a dedicated type for a field only containing settled capsule
elements harder.

Ideally, `Arbitrary::arbitrary` and `Arbitrary::shrink` would already
produce fields which are in themself consistent. This change starts of a
transition to this approach by introducing a utility for making capsules
consistent.