field: migrate tests::MovingField to new consistency model
For tests involving prepared fields, we generate a set of
`RandomCapsules`. Currently, we'd accept inconsistencies in those sets
during generation and filter them out during the transformation into an
actual `field::MovingField`. This approach turned out to be non-ideal.
Ideally, `Arbitrary::arbitrary` and `Arbitrary::shrink` would already
produce fields which are in themself consistent.

This change achieves this by migrating to using a recently introduced
utility for ensuring consistency.