field: add util for generating StaticFields without unsupported elems
We already have a utility for generating `StaticField`s. However, the
generated fields may contain capsules which are not supported by another
capsule or virus. While we plan to introduce tests which would rely on
all capsules having some support we don't want to bring this assumption
into existing tests.

We therefore decided to introduce an additional generation helper.