field: unsettle ex-partners of eliminated capsules if necessary
Up until now, we'd perform unsettlement based only on eliminated
capsules. We'd neglect the possibilities of partners of those capsules
to loose suppport due to the event, which resulted in behaviour
violating the specification.

This change makes `unsettle_capsules` take ex-partners into account. In
addition, we took the opportunity to migrate to a heap for the worklist,
which should bring with it some (very) minor performance gains.

Fix-phase: testing