field: fix determination of lowest row containing unsettled capsules
... in `unsettle_elements`.

This one caused a lot of spurious failures in a test of
`game::round::Actor::tick` which is yet to be comitted.

This error was hard to find, even though we had a suspition that the
failures stemmed from some logic error in `unsettle_elements` early on.
And we ended up extending our testing on field as a result, albeit those
tests wouldn't uncover the issue.

Fix-phase: testing