field: omit always-true condition in unsettle_elements
`worklist` used to be a `Vec`. Hence, positions popped from the worklist
could be in any order. If we wanted to determine the lowest row
containing a capsule we unsettled, we'd need to compare the row to the
current lowest. Incidentally, this very comparison turned out to be
erroneous, which we fixed in

        10274ac60f607e9919e6dad88cc187bd62d68558

However, in

        955bc2726fb63d5bb3dca26e34db17214929227c

we already had migrated to a heap for the worklist. Since, during the
iteration, we'd only insert positions above the currently processed one,
we are guaranteed to never encounter positions in rows lower than the
one we are currently processing. Thus, once set, `lowest_unsettled` will
not change. The comparison is thus unnecessary.

Normally, we'd refrain from introducing such changes during the testing
phase. However, we justify doing so anyway because it amounts to an
alternative fix.