game: change layout of left column during round to fit contents
When writing `round::serve` we erred on the number of rows the play
field would occupy and assigned more rows to the indication text field
than there were left in the column. And we swapped the number of rows
and columns for the field construction.

This change squeeses the indicator into the column by placing it just
below the play field, without any padding. We also needed to reduce it
to the two rows required by the longest message we may need to display
during a round.